Extracting StyleRemovingTextWatcher from BrowserActivity for reduced complexity
